refactor tuple data structures into separate files
